db_project/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap4==2024.1
decorator==5.1.1
Django==5.0.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==24.3
django-geocoder==0.2.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.4.0
geocoder==1.15.1
idna==3.7
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
ratelim==0.1.6
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
